feat: output latest wanted version according to semver caret range
The output of the application will now also show the latest available
version, according to the image's current version semver-caret-range.
This is especially helpful, if an installed image cannot be updated to
the next major version but still wants to be notified about
patch/minor updates.

BREAKING CHANGE: Output schema of the application changed -> Column
"Upgrade Type" was removed